pyproject.toml
[project]
name = "content-workflow-agent"
version = "1.0.0"
description = "Convert blog posts to platform-specific social media content"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "openai>=1.52.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "wikipedia>=1.4.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "numpy>=2.1.0",
    "scikit-learn>=1.5.0",
    "tiktoken>=0.7.0",
    "pytest>=8.3.0",
    "streamlit>=1.48.1",
    "requests>=2.32.5",
    "ddgs>=9.5.4",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]
